+++ src/main/java/me/alidg/errors/WebErrorHandlers.java
@@ -1,7 +1,7 @@
 package me.alidg.errors;
 
 import me.alidg.errors.HttpError.CodedMessage;
-import me.alidg.errors.impl.LastResortWebErrorHandler;
+import me.alidg.errors.handlers.LastResortWebErrorHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.MessageSource;
