+++ src/test/java/me/alidg/errors/WebErrorHandlersIT.java
@@ -6,7 +6,7 @@ import me.alidg.errors.HttpError.CodedMessage;
 import me.alidg.errors.annotation.ExceptionMapping;
 import me.alidg.errors.annotation.ExposeAsArg;
 import me.alidg.errors.conf.ErrorsAutoConfiguration;
-import me.alidg.errors.impl.LastResortWebErrorHandler;
+import me.alidg.errors.handlers.LastResortWebErrorHandler;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
