+++ src/test/java/me/alidg/errors/conf/ErrorsAutoConfigurationIT.java
@@ -3,7 +3,7 @@ package me.alidg.errors.conf;
 import me.alidg.errors.HandledException;
 import me.alidg.errors.WebErrorHandler;
 import me.alidg.errors.WebErrorHandlers;
-import me.alidg.errors.impl.*;
+import me.alidg.errors.handlers.*;
 import org.junit.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.TestConfiguration;
