+++ src/test/java/me/alidg/errors/handlers/AnnotatedWebErrorHandlerTest.java
@@ -1,4 +1,4 @@
-package me.alidg.errors.impl;
+package me.alidg.errors.handlers;
 
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
