+++ src/test/java/me/alidg/errors/handlers/SpringValidationWebErrorHandlerTest.java
@@ -1,4 +1,4 @@
-package me.alidg.errors.impl;
+package me.alidg.errors.handlers;
 
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
@@ -23,7 +23,7 @@ import java.util.*;
 import static java.util.Arrays.asList;
 import static java.util.Collections.*;
 import static me.alidg.Params.p;
-import static me.alidg.errors.impl.SpringValidationWebErrorHandlerTest.TBV.tbv;
+import static me.alidg.errors.handlers.SpringValidationWebErrorHandlerTest.TBV.tbv;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
