+++ src/test/java/me/alidg/errors/mvc/ErrorsControllerAdviceIT.java
@@ -7,7 +7,7 @@ import me.alidg.errors.HttpError.CodedMessage;
 import me.alidg.errors.annotation.ExceptionMapping;
 import me.alidg.errors.annotation.ExposeAsArg;
 import me.alidg.errors.conf.ErrorsAutoConfiguration;
-import me.alidg.errors.impl.LastResortWebErrorHandler;
+import me.alidg.errors.handlers.LastResortWebErrorHandler;
 import org.hamcrest.Matchers;
 import org.junit.After;
 import org.junit.Before;
@@ -57,9 +57,9 @@ import java.util.Locale;
 import java.util.stream.Stream;
 
 import static me.alidg.Params.p;
-import static me.alidg.errors.impl.SpringMvcWebErrorHandler.*;
-import static me.alidg.errors.impl.SpringSecurityWebErrorHandler.ACCESS_DENIED;
-import static me.alidg.errors.impl.SpringSecurityWebErrorHandler.AUTH_REQUIRED;
+import static me.alidg.errors.handlers.SpringMvcWebErrorHandler.*;
+import static me.alidg.errors.handlers.SpringSecurityWebErrorHandler.ACCESS_DENIED;
+import static me.alidg.errors.handlers.SpringSecurityWebErrorHandler.AUTH_REQUIRED;
 import static me.alidg.errors.mvc.ErrorsControllerAdviceIT.Dto.dto;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
